% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analyze_cluster_coherence}
\alias{analyze_cluster_coherence}
\title{Comparer les variables intra-cluster}
\usage{
analyze_cluster_coherence(X, groupes)
}
\arguments{
\item{X}{Data frame de données (variables numériques)}

\item{groupes}{Vecteur des groupes de variables}
}
\value{
Data frame avec statistiques par cluster
}
\description{
Analyse la cohérence des variables au sein de chaque cluster
}
