% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Vérifier la présence de valeurs manquantes}
\usage{
check_missing_values(X, verbose = TRUE)
}
\arguments{
\item{X}{Data frame ou matrice}

\item{verbose}{Booléen, afficher les détails (défaut: TRUE)}
}
\value{
Liste avec statistiques sur les NA
}
\description{
Analyse les valeurs manquantes dans le dataset
}
