% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAnalysis.R
\name{ClusterAnalysis}
\alias{ClusterAnalysis}
\title{Parent Class for Cluster Analysis}
\description{
\code{ClusterAnalysis} is an R6 parent class that defines the base architecture
for all clustering algorithms in the RollerClustR package. It implements
the Template Method pattern to ensure a consistent interface across
different clustering methods.
}
\details{
This class should not be instantiated directly. It serves as a base for
child classes: \link{VAR_CAH}, \link{VAR_KMEANS} and \link{TandemVarClust}.
\subsection{Public Methods}{
\itemize{
\item \code{fit(X)} : Fits the model to the data
\item \code{summary()} : Displays a summary of the fitted model
\item \code{predict(newdata)} : Predicts groups for new data
}
}

\subsection{Active Bindings}{
\itemize{
\item \code{K} : Number of clusters (read/write)
\item \code{Groupes} : Vector of group assignments
}
}

\subsection{Inheritance Contracts}{

Child classes must implement the following private methods:
\itemize{
\item \code{do_fit(X)} : Specific fitting logic
\item \code{do_refit_with_k(new_k)} : Refitting with a new K
\item \code{do_predict(newdata)} : Prediction for new data
\item \code{do_summary()} : Display of specific summary
}
}
}
\examples{
# This class should not be used directly
# Use child classes instead:
\dontrun{
model <- VAR_CAH$new(K = 3)
model$fit(iris[, 1:4])
model$summary()
}


## ------------------------------------------------
## Method `ClusterAnalysis$fit`
## ------------------------------------------------

\dontrun{
model <- VAR_CAH$new(K = 3)
model$fit(iris[, 1:4])
}

## ------------------------------------------------
## Method `ClusterAnalysis$summary`
## ------------------------------------------------

\dontrun{
model <- VAR_CAH$new(K = 3)
model$fit(iris[, 1:4])
model$summary()
}

## ------------------------------------------------
## Method `ClusterAnalysis$predict`
## ------------------------------------------------

\dontrun{
model <- VAR_CAH$new(K = 3)
model$fit(iris[1:100, 1:4])
predictions <- model$predict(iris[101:150, 1:4])
}
}
\seealso{
\link{VAR_CAH}, \link{VAR_KMEANS}, \link{TandemVarClust}, \code{\link[=roller_clust]{roller_clust()}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{K}}{Number of clusters (read/write).
Writing triggers a model refitting.}

\item{\code{Groupes}}{Vector of group assignments for each
observation or variable (depending on clustering method).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusterAnalysis-new}{\code{ClusterAnalysis$new()}}
\item \href{#method-ClusterAnalysis-fit}{\code{ClusterAnalysis$fit()}}
\item \href{#method-ClusterAnalysis-summary}{\code{ClusterAnalysis$summary()}}
\item \href{#method-ClusterAnalysis-predict}{\code{ClusterAnalysis$predict()}}
\item \href{#method-ClusterAnalysis-clone}{\code{ClusterAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterAnalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterAnalysis-new}{}}}
\subsection{Method \code{new()}}{
Abstract method for fitting


Abstract method for refitting with new K


Abstract method for prediction


Abstract method for summary


Constructor of parent class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterAnalysis$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to child classes}

\item{\code{X}}{Data frame or data matrix}

\item{\code{new_k}}{New number of clusters}

\item{\code{newdata}}{New data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterAnalysis-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterAnalysis-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the model to the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterAnalysis$fit(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame or matrix of training data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object itself (invisibly) to allow chaining
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
model <- VAR_CAH$new(K = 3)
model$fit(iris[, 1:4])
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterAnalysis-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterAnalysis-summary}{}}}
\subsection{Method \code{summary()}}{
Displays a summary of the fitted model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterAnalysis$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object itself (invisibly)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
model <- VAR_CAH$new(K = 3)
model$fit(iris[, 1:4])
model$summary()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterAnalysis-predict"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterAnalysis-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts groups for new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterAnalysis$predict(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{New data (same structure as training X)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of group predictions
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
model <- VAR_CAH$new(K = 3)
model$fit(iris[1:100, 1:4])
predictions <- model$predict(iris[101:150, 1:4])
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusterAnalysis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusterAnalysis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusterAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
