% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ClusteringComparator}
\alias{ClusteringComparator}
\title{ClusteringComparator}
\description{
Outil pour comparer les performances de differents objets de clustering
}
\details{
Outil pour comparer les performances de differents objets de clustering
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusteringComparator-new}{\code{ClusteringComparator$new()}}
\item \href{#method-ClusteringComparator-compare_inertia}{\code{ClusteringComparator$compare_inertia()}}
\item \href{#method-ClusteringComparator-get_best_model}{\code{ClusteringComparator$get_best_model()}}
\item \href{#method-ClusteringComparator-clone}{\code{ClusteringComparator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringComparator-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringComparator-new}{}}}
\subsection{Method \code{new()}}{
Initialiser le Comparator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringComparator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringComparator-compare_inertia"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringComparator-compare_inertia}{}}}
\subsection{Method \code{compare_inertia()}}{
Comparer les inerties de plusieurs modeles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringComparator$compare_inertia(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Objets de clustering a comparer (ajustes avec fit())}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un data frame de comparaison
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringComparator-get_best_model"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringComparator-get_best_model}{}}}
\subsection{Method \code{get_best_model()}}{
Identifier le meilleur modele selon un critere
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringComparator$get_best_model(comparison_df, criterion = "Pct_Expliquee")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comparison_df}}{Data frame retourne par compare_inertia()}

\item{\code{criterion}}{Critere de selection (defaut: "Pct_Expliquee")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Indice du meilleur modele
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringComparator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringComparator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringComparator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
