% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ClusteringEvaluator}
\alias{ClusteringEvaluator}
\title{ClusteringEvaluator}
\description{
Outil pour evaluer le nombre optimal de clusters de variables
}
\details{
Outil pour evaluer le nombre optimal de clusters de variables
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusteringEvaluator-new}{\code{ClusteringEvaluator$new()}}
\item \href{#method-ClusteringEvaluator-elbow_method}{\code{ClusteringEvaluator$elbow_method()}}
\item \href{#method-ClusteringEvaluator-evaluate_k_range}{\code{ClusteringEvaluator$evaluate_k_range()}}
\item \href{#method-ClusteringEvaluator-plot_evaluation}{\code{ClusteringEvaluator$plot_evaluation()}}
\item \href{#method-ClusteringEvaluator-get_best_k}{\code{ClusteringEvaluator$get_best_k()}}
\item \href{#method-ClusteringEvaluator-clone}{\code{ClusteringEvaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-new}{}}}
\subsection{Method \code{new()}}{
Initialiser l'Evaluator avec les donnees (optionnel)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$new(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame de donnees (optionnel)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-elbow_method"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-elbow_method}{}}}
\subsection{Method \code{elbow_method()}}{
Methode du coude (Elbow Method)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$elbow_method(X, max_k = 10, method = "var_cah", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame de donnees}

\item{\code{max_k}}{Nombre maximal de clusters a tester (defaut: 10)}

\item{\code{method}}{Methode de clustering (defaut: "var_cah")}

\item{\code{...}}{Parametres additionnels pour la methode}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Liste contenant le graphique (ggplot) et les donnees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-evaluate_k_range"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-evaluate_k_range}{}}}
\subsection{Method \code{evaluate_k_range()}}{
Evaluer une plage de valeurs de k
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$evaluate_k_range(
  X,
  k_range = 2:10,
  method = "var_cah",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame de donnees}

\item{\code{k_range}}{Vecteur de valeurs de k a tester (defaut: 2:10)}

\item{\code{method}}{Methode de clustering (defaut: "var_cah")}

\item{\code{...}}{Parametres additionnels}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame avec les resultats pour chaque k
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-plot_evaluation"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-plot_evaluation}{}}}
\subsection{Method \code{plot_evaluation()}}{
Visualiser l'evaluation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$plot_evaluation(resultats, criterion = "inertie_expliquee")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resultats}}{Data frame de resultats (de evaluate_k_range)}

\item{\code{criterion}}{Critere a visualiser (defaut: "inertie_expliquee")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-get_best_k"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-get_best_k}{}}}
\subsection{Method \code{get_best_k()}}{
Obtenir le meilleur k selon un critere
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$get_best_k(resultats, criterion = "inertie_expliquee")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resultats}}{Data frame de resultats}

\item{\code{criterion}}{Critere (defaut: "inertie_expliquee")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Valeur de k recommandee
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
