% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ClusteringHelper}
\alias{ClusteringHelper}
\title{ClusteringHelper}
\description{
Fonctions utilitaires pour le package de clustering de variables
}
\details{
Fonctions utilitaires pour le package de clustering de variables
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusteringHelper-new}{\code{ClusteringHelper$new()}}
\item \href{#method-ClusteringHelper-generate_report}{\code{ClusteringHelper$generate_report()}}
\item \href{#method-ClusteringHelper-show_methods}{\code{ClusteringHelper$show_methods()}}
\item \href{#method-ClusteringHelper-export_results}{\code{ClusteringHelper$export_results()}}
\item \href{#method-ClusteringHelper-clone}{\code{ClusteringHelper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringHelper-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringHelper-new}{}}}
\subsection{Method \code{new()}}{
Initialiser le Helper
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringHelper$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringHelper-generate_report"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringHelper-generate_report}{}}}
\subsection{Method \code{generate_report()}}{
Genere un rapport textuel synthetique du clustering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringHelper$generate_report(objet_clustering, file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objet_clustering}}{Objet de clustering ajuste (VAR_CAH, VAR_KMEANS, ou TandemVarClust)}

\item{\code{file}}{Chemin du fichier ou ecrire le rapport (NULL pour affichage console)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisiblement, l'objet lui-meme. Affiche le rapport.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringHelper-show_methods"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringHelper-show_methods}{}}}
\subsection{Method \code{show_methods()}}{
Affiche les methodes disponibles dans le package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringHelper$show_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringHelper-export_results"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringHelper-export_results}{}}}
\subsection{Method \code{export_results()}}{
Exporter les resultats du clustering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringHelper$export_results(
  objet_clustering,
  file = "clustering_results.csv"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objet_clustering}}{Objet de clustering ajuste}

\item{\code{file}}{Chemin du fichier CSV de sortie}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringHelper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringHelper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringHelper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
