% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_clustering_results}
\alias{export_clustering_results}
\title{Exporter les résultats du clustering}
\usage{
export_clustering_results(clustering_object, file = "clustering_results.csv")
}
\arguments{
\item{clustering_object}{Objet de clustering ajusté}

\item{file}{Chemin du fichier de sortie (défaut: "clustering_results.csv")}
}
\value{
Data frame des résultats (invisible)
}
\description{
Exporte les groupes de variables dans un fichier CSV
}
