% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cluster_quality}
\alias{plot_cluster_quality}
\title{Visualiser la qualité du clustering}
\usage{
plot_cluster_quality(inertia_data)
}
\arguments{
\item{inertia_data}{Data frame avec colonnes 'k' et 'inertie_expliquee'}
}
\description{
Graphique de l'inertie expliquée en fonction du nombre de clusters
}
