% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TandemVarClust.R
\name{print.TandemVarClust_predict}
\alias{print.TandemVarClust_predict}
\title{Print method for TandemVarClust prediction results}
\usage{
\method{print}{TandemVarClust_predict}(x, ...)
}
\arguments{
\item{x}{A TandemVarClust_predict object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns x
}
\description{
Print method for TandemVarClust prediction results
}
