% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_clustering}
\alias{summarize_clustering}
\title{Créer un résumé textuel du clustering}
\usage{
summarize_clustering(clustering_object)
}
\arguments{
\item{clustering_object}{Objet de clustering ajusté}
}
\value{
Caractère avec le résumé
}
\description{
Génère un résumé structuré des résultats
}
