% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_CAH.R
\name{VAR_CAH}
\alias{VAR_CAH}
\title{Clustering Ascendant Hiérarchique de Variables (VAR_CAH)}
\value{
Un objet R6 de classe \code{VAR_CAH}
}
\description{
\code{VAR_CAH} implémente un algorithme de clustering ascendant hiérarchique (CAH)
pour le regroupement de variables numériques. L'algorithme utilise la matrice
de corrélation pour calculer les distances entre variables et construit un
dendrogramme permettant d'identifier des groupes de variables similaires.
}
\details{
\subsection{Principe de l'algorithme}{
\enumerate{
\item Calcul de la matrice de corrélation entre toutes les variables
\item Transformation en matrice de distance : \verb{d = 1 - |cor|}
\item Construction de l'arbre hiérarchique (méthode complete linkage)
\item Découpe de l'arbre pour obtenir K clusters
\item Calcul de variables synthétiques par ACP pour chaque cluster
}
}

\subsection{Variables synthétiques}{

Pour chaque cluster de variables, une variable synthétique est calculée comme
la première composante principale (PC1) des variables du cluster. Cette
synthèse maximise la variance expliquée et fournit une représentation optimale
du cluster.
}

\subsection{Métriques de qualité}{
\itemize{
\item \strong{Homogénéité} : Corrélation moyenne intra-cluster
\item \strong{Qualité des clusters} : Variance expliquée par PC1
\item \strong{Corrélations} : Matrice de corrélation complète
}
}
}
\section{Méthodes publiques}{

\describe{
\item{\verb{$new(...)}}{Constructeur de la classe}
\item{\verb{$fit(X)}}{Ajuste le modèle sur les données X}
\item{\verb{$summary()}}{Affiche un résumé détaillé du clustering}
\item{\verb{$predict(newdata)}}{Non implémenté pour VAR_CAH (clustering de variables)}
}
}

\section{Active bindings}{

\describe{
\item{\verb{$K}}{Nombre de clusters (lecture/écriture)}
\item{\verb{$Groupes}}{Vecteur nommé des affectations de variables aux clusters}
}
}

\examples{
# Exemple simple avec le jeu de données iris
library(RollerClustR)
data(iris)

# Créer et ajuster le modèle
model <- VAR_CAH$new(K = 2, scale = TRUE)
model$fit(iris[, 1:4])

# Afficher le résumé
model$summary()

# Accéder aux groupes de variables
groups <- model$Groupes
print(groups)

# Modifier le nombre de clusters
model$K <- 3
model$summary()

}
\seealso{
\link{ClusterAnalysis}, \link{VARCLUS}, \link{KmodesVarClust}, \code{\link[=roller_clust]{roller_clust()}}
}
\section{Super class}{
\code{\link[RollerClustR:ClusterAnalysis]{RollerClustR::ClusterAnalysis}} -> \code{VAR_CAH}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Groupes}}{Groupes des variables (surcharge avec noms de variables)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VAR_CAH-new}{\code{VAR_CAH$new()}}
\item \href{#method-VAR_CAH-get_cluster_variables}{\code{VAR_CAH$get_cluster_variables()}}
\item \href{#method-VAR_CAH-get_representative_variable}{\code{VAR_CAH$get_representative_variable()}}
\item \href{#method-VAR_CAH-get_tree}{\code{VAR_CAH$get_tree()}}
\item \href{#method-VAR_CAH-inertie}{\code{VAR_CAH$inertie()}}
\item \href{#method-VAR_CAH-clone}{\code{VAR_CAH$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="fit"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-fit'><code>RollerClustR::ClusterAnalysis$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="predict"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-predict'><code>RollerClustR::ClusterAnalysis$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="summary"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-summary'><code>RollerClustR::ClusterAnalysis$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-new"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-new}{}}}
\subsection{Method \code{new()}}{
Calcule la variable synthétique (PC1) pour un cluster numérique


Méthode do_fit : ajuste l'algorithme CAH sur les variables.


Méthode interne pour refaire l'ajustement si K est modifié


Méthode predict (non implémentée pour clustering de variables)


Évaluation des clusters


Méthode summary privée


Constructeur de la classe VAR_CAH
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$new(scale = TRUE, K = 2, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{Booléen indiquant si les variables doivent être centrées-réduites (défaut : TRUE)}

\item{\code{K}}{Nombre de clusters souhaités (défaut : 2)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-get_cluster_variables"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-get_cluster_variables}{}}}
\subsection{Method \code{get_cluster_variables()}}{
Obtenir les variables appartenant à un cluster donné
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$get_cluster_variables(cluster_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-get_representative_variable"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-get_representative_variable}{}}}
\subsection{Method \code{get_representative_variable()}}{
Obtenir la variable la plus représentative d'un cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$get_representative_variable(cluster_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-get_tree"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-get_tree}{}}}
\subsection{Method \code{get_tree()}}{
Obtenir l'arbre hiérarchique des variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$get_tree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-inertie"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-inertie}{}}}
\subsection{Method \code{inertie()}}{
Calculer les inerties du clustering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$inertie()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Cette méthode calcule l'inertie totale, intra-cluster et inter-cluster
basée sur les corrélations des variables avec leur composante principale.
}

\subsection{Returns}{
Liste avec les éléments suivants:
\itemize{
\item totale: Variance totale des corrélations
\item intra: Inertie intra-cluster (variance moyenne dans les clusters)
\item inter: Inertie inter-cluster (totale - intra)
\item pct_expliquee: Pourcentage de variance expliquée par le clustering
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
