% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARCLUS.R
\name{VARCLUS}
\alias{VARCLUS}
\title{Clustering Descendant Hierarchique de Variables (VARCLUS)}
\description{
\code{VARCLUS} implemente un algorithme de clustering descendant (divisif) pour le
regroupement de variables. L'algorithme procede par division recursive des
groupes de variables en utilisant l'analyse en composantes principales (ACP)
et la rotation Varimax.
}
\details{
\subsection{Principe de l'algorithme}{

VARCLUS utilise une approche descendante (top-down) :
\enumerate{
\item Initialisation : toutes les variables dans un cluster
\item Division recursive basee sur ACP et rotation Varimax
\item Critere d'arret : λ₂ < seuil ou < 3 variables
}
}
}
\section{Super class}{
\code{\link[RollerClustR:ClusterAnalysis]{RollerClustR::ClusterAnalysis}} -> \code{VARCLUS}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{K}}{Nombre de clusters (lecture seule pour VARCLUS)}

\item{\code{Groupes}}{Groupes des variables (lecture seule)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VARCLUS-new}{\code{VARCLUS$new()}}
\item \href{#method-VARCLUS-cut_tree}{\code{VARCLUS$cut_tree()}}
\item \href{#method-VARCLUS-inertie}{\code{VARCLUS$inertie()}}
\item \href{#method-VARCLUS-clone}{\code{VARCLUS$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="fit"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-fit'><code>RollerClustR::ClusterAnalysis$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="predict"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-predict'><code>RollerClustR::ClusterAnalysis$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="summary"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-summary'><code>RollerClustR::ClusterAnalysis$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VARCLUS-new"></a>}}
\if{latex}{\out{\hypertarget{method-VARCLUS-new}{}}}
\subsection{Method \code{new()}}{
Methode do_fit (appelee par fit())


Methode do_refit_with_k (non applicable pour VARCLUS)


Prédit le cluster d'appartenance pour de nouvelles variables


Methode do_summary (appelee par summary())


Constructeur de la classe VARCLUS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VARCLUS$new(
  K = NULL,
  stop_eigenvalue = 1,
  distance_metric = "correlation",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{Nombre de clusters (ignore - determine automatiquement)}

\item{\code{stop_eigenvalue}}{Seuil d'arret pour λ₂ (defaut: 1.0)}

\item{\code{distance_metric}}{Metrique de distance (defaut: "correlation")}

\item{\code{...}}{Autres parametres (pour compatibilite)}

\item{\code{newdata}}{Data frame ou vecteur contenant les nouvelles variables}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Liste contenant les assignations et les scores de similarité
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VARCLUS-cut_tree"></a>}}
\if{latex}{\out{\hypertarget{method-VARCLUS-cut_tree}{}}}
\subsection{Method \code{cut_tree()}}{
Couper l'arbre pour obtenir k clusters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VARCLUS$cut_tree(k)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{Nombre de clusters souhaite}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vecteur des groupes de variables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VARCLUS-inertie"></a>}}
\if{latex}{\out{\hypertarget{method-VARCLUS-inertie}{}}}
\subsection{Method \code{inertie()}}{
Calculer les inerties du clustering
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VARCLUS$inertie()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Cette methode calcule l'inertie basee sur les valeurs propres
des ACP realisees sur chaque cluster de variables.
}

\subsection{Returns}{
Liste avec les elements suivants:
\itemize{
\item totale: Inertie totale (nombre de variables, maximum theorique)
\item intra: Inertie intra-cluster (variance non expliquee)
\item inter: Inertie inter-cluster (variance expliquee par le clustering)
\item pct_expliquee: Pourcentage de variance expliquee
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VARCLUS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VARCLUS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VARCLUS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
