% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ClusteringEvaluator}
\alias{ClusteringEvaluator}
\title{ClusteringEvaluator}
\description{
Outil pour evaluer le nombre optimal de clusters (k)
}
\details{
Outil pour evaluer le nombre optimal de clusters
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusteringEvaluator-new}{\code{ClusteringEvaluator$new()}}
\item \href{#method-ClusteringEvaluator-evaluate_k_var_cah}{\code{ClusteringEvaluator$evaluate_k_var_cah()}}
\item \href{#method-ClusteringEvaluator-evaluate_k_var_kmeans}{\code{ClusteringEvaluator$evaluate_k_var_kmeans()}}
\item \href{#method-ClusteringEvaluator-plot_evaluation}{\code{ClusteringEvaluator$plot_evaluation()}}
\item \href{#method-ClusteringEvaluator-get_best_k}{\code{ClusteringEvaluator$get_best_k()}}
\item \href{#method-ClusteringEvaluator-clone}{\code{ClusteringEvaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-new}{}}}
\subsection{Method \code{new()}}{
Initialiser l'Evaluator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-evaluate_k_var_cah"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-evaluate_k_var_cah}{}}}
\subsection{Method \code{evaluate_k_var_cah()}}{
Evaluer plusieurs valeurs de k pour VAR_CAH
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$evaluate_k_var_cah(X, k_min = 2, k_max = 10, scale = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame ou matrice de donnees}

\item{\code{k_min}}{Nombre minimum de clusters (defaut: 2)}

\item{\code{k_max}}{Nombre maximum de clusters (defaut: 10)}

\item{\code{scale}}{Booleen, standardiser les donnees (defaut: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame avec k et les metriques associees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-evaluate_k_var_kmeans"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-evaluate_k_var_kmeans}{}}}
\subsection{Method \code{evaluate_k_var_kmeans()}}{
Evaluer plusieurs valeurs de k pour VAR_KMEANS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$evaluate_k_var_kmeans(
  X,
  k_min = 2,
  k_max = 10,
  n_init = 10,
  scale = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame ou matrice de donnees}

\item{\code{k_min}}{Nombre minimum de clusters (defaut: 2)}

\item{\code{k_max}}{Nombre maximum de clusters (defaut: 10)}

\item{\code{n_init}}{Nombre d'initialisations (defaut: 10)}

\item{\code{scale}}{Booleen, standardiser les donnees (defaut: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame avec k et les metriques associees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-plot_evaluation"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-plot_evaluation}{}}}
\subsection{Method \code{plot_evaluation()}}{
Visualiser les resultats d'evaluation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$plot_evaluation(resultats, criterion = "inertie_expliquee")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resultats}}{Data frame retourne par evaluate_k_*}

\item{\code{criterion}}{Critere a visualiser (defaut: "inertie_expliquee")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-get_best_k"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-get_best_k}{}}}
\subsection{Method \code{get_best_k()}}{
Obtenir le meilleur k selon un critere
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$get_best_k(resultats, criterion = "inertie_expliquee")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resultats}}{Data frame de resultats}

\item{\code{criterion}}{Critere (defaut: "inertie_expliquee")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Valeur de k recommandee
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringEvaluator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringEvaluator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringEvaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
