% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ClusteringFactory}
\alias{ClusteringFactory}
\title{ClusteringFactory}
\description{
Factory pour creer et initialiser les objets de clustering de variables
}
\details{
Factory pour créer et initialiser les objets de clustering de variables
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClusteringFactory-new}{\code{ClusteringFactory$new()}}
\item \href{#method-ClusteringFactory-create_var_cah}{\code{ClusteringFactory$create_var_cah()}}
\item \href{#method-ClusteringFactory-create_kmodes_varclust}{\code{ClusteringFactory$create_kmodes_varclust()}}
\item \href{#method-ClusteringFactory-create_varclus}{\code{ClusteringFactory$create_varclus()}}
\item \href{#method-ClusteringFactory-clone}{\code{ClusteringFactory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringFactory-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringFactory-new}{}}}
\subsection{Method \code{new()}}{
Initialiser la Factory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringFactory$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringFactory-create_var_cah"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringFactory-create_var_cah}{}}}
\subsection{Method \code{create_var_cah()}}{
Creer et ajuster un objet VAR_CAH (Clustering de Variables - CAH)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringFactory$create_var_cah(X, k = 3, scale = TRUE, fit_now = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame ou matrice de donnees}

\item{\code{k}}{Nombre de clusters (defaut: 3)}

\item{\code{scale}}{Booleen, standardiser les donnees (defaut: TRUE)}

\item{\code{fit_now}}{Booleen, ajuster le modele immediatement (defaut: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Objet VAR_CAH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringFactory-create_kmodes_varclust"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringFactory-create_kmodes_varclust}{}}}
\subsection{Method \code{create_kmodes_varclust()}}{
Creer et ajuster un objet KmodesVarClust
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringFactory$create_kmodes_varclust(
  X,
  k = 3,
  max_iter = 100,
  n_bins = 5,
  fit_now = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame ou matrice de donnees}

\item{\code{k}}{Nombre de clusters (defaut: 3)}

\item{\code{max_iter}}{Nombre maximum d'iterations (defaut: 100)}

\item{\code{n_bins}}{Nombre d'intervalles pour discretisation (defaut: 5)}

\item{\code{fit_now}}{Booleen, ajuster le modele immediatement (defaut: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Objet KmodesVarClust
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringFactory-create_varclus"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringFactory-create_varclus}{}}}
\subsection{Method \code{create_varclus()}}{
Creer et ajuster un objet VARCLUS (Clustering de Variables - Descendant)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringFactory$create_varclus(
  X,
  stop_eigenvalue = 1,
  distance_metric = "correlation",
  fit_now = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Data frame ou matrice de donnees}

\item{\code{stop_eigenvalue}}{Seuil d'arret pour lambda2 (defaut: 1.0)}

\item{\code{distance_metric}}{Metrique de distance (defaut: "correlation")}

\item{\code{fit_now}}{Booleen, ajuster le modele immediatement (defaut: TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Objet VARCLUS
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClusteringFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClusteringFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClusteringFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
