% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_CAH.R
\name{VAR_CAH}
\alias{VAR_CAH}
\title{Hierarchical Agglomerative Clustering of Variables (VAR_CAH)}
\value{
An R6 object of class \code{VAR_CAH}
}
\description{
\code{VAR_CAH} implements a hierarchical agglomerative clustering (HAC) algorithm
for grouping numeric variables. The algorithm uses the correlation matrix
to compute distances between variables and builds a dendrogram to identify
groups of similar variables.
}
\details{
\subsection{Algorithm Principle}{
\enumerate{
\item Compute correlation matrix between all variables
\item Transform into distance matrix: \verb{d = 1 - |cor|}
\item Build hierarchical tree (complete linkage method)
\item Cut tree to obtain K clusters
\item Compute synthetic variables by PCA for each cluster
}
}

\subsection{Synthetic Variables}{

For each variable cluster, a synthetic variable is computed as the first
principal component (PC1) of the cluster variables. This synthesis maximizes
explained variance and provides an optimal representation of the cluster.
}

\subsection{Quality Metrics}{
\itemize{
\item \strong{Homogeneity}: Average intra-cluster correlation
\item \strong{Cluster Quality}: Variance explained by PC1
\item \strong{Correlations}: Complete correlation matrix
}
}
}
\section{Public methods}{

\describe{
\item{\verb{$new(...)}}{Class constructor}
\item{\verb{$fit(X)}}{Fits the model on data X}
\item{\verb{$summary()}}{Displays a detailed clustering summary}
\item{\verb{$predict(newdata)}}{Not implemented for VAR_CAH (variable clustering)}
}
}

\section{Active bindings}{

\describe{
\item{\verb{$K}}{Number of clusters (read/write)}
\item{\verb{$Groupes}}{Named vector of variable assignments to clusters}
}
}

\examples{
# Simple example with iris dataset
library(RollerClustR)
data(iris)

# Create and fit the model
model <- VAR_CAH$new(K = 2, scale = TRUE)
model$fit(iris[, 1:4])

# Display summary
model$summary()

# Access variable groups
groups <- model$Groupes
print(groups)

# Modify number of clusters
model$K <- 3
model$summary()

}
\seealso{
\link{ClusterAnalysis}, \link{VARCLUS}, \code{TandemVarClust}, \code{\link[=roller_clust]{roller_clust()}}
}
\section{Super class}{
\code{\link[RollerClustR:ClusterAnalysis]{RollerClustR::ClusterAnalysis}} -> \code{VAR_CAH}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Groupes}}{Variable groups (override with variable names)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VAR_CAH-new}{\code{VAR_CAH$new()}}
\item \href{#method-VAR_CAH-get_cluster_variables}{\code{VAR_CAH$get_cluster_variables()}}
\item \href{#method-VAR_CAH-get_representative_variable}{\code{VAR_CAH$get_representative_variable()}}
\item \href{#method-VAR_CAH-get_tree}{\code{VAR_CAH$get_tree()}}
\item \href{#method-VAR_CAH-inertie}{\code{VAR_CAH$inertie()}}
\item \href{#method-VAR_CAH-clone}{\code{VAR_CAH$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="fit"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-fit'><code>RollerClustR::ClusterAnalysis$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="predict"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-predict'><code>RollerClustR::ClusterAnalysis$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RollerClustR" data-topic="ClusterAnalysis" data-id="summary"><a href='../../RollerClustR/html/ClusterAnalysis.html#method-ClusterAnalysis-summary'><code>RollerClustR::ClusterAnalysis$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-new"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-new}{}}}
\subsection{Method \code{new()}}{
Computes synthetic variable (PC1) for a numeric cluster


do_fit method: fits HAC algorithm on variables


Internal method to refit if K is modified


Predict cluster membership for new variables


Cluster evaluation


Private summary method


VAR_CAH class constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$new(scale = TRUE, K = 2, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{Boolean indicating if variables should be centered and scaled (default: TRUE)}

\item{\code{K}}{Desired number of clusters (default: 2)}

\item{\code{...}}{Additional arguments (for compatibility with parent class)}

\item{\code{newdata}}{Data frame or vector containing new variables}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List containing assignments and similarity scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-get_cluster_variables"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-get_cluster_variables}{}}}
\subsection{Method \code{get_cluster_variables()}}{
Get variables belonging to a given cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$get_cluster_variables(cluster_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cluster_id}}{Cluster identifier (integer between 1 and K)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-get_representative_variable"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-get_representative_variable}{}}}
\subsection{Method \code{get_representative_variable()}}{
Get the most representative variable of a cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$get_representative_variable(cluster_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cluster_id}}{Cluster identifier (integer between 1 and K)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-get_tree"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-get_tree}{}}}
\subsection{Method \code{get_tree()}}{
Get the hierarchical tree of variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$get_tree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-inertie"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-inertie}{}}}
\subsection{Method \code{inertie()}}{
Compute clustering inertias
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$inertie()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method computes total, within-cluster and between-cluster inertia
based on correlations of variables with their principal component.
}

\subsection{Returns}{
List with following elements:
\itemize{
\item totale: Total variance of correlations
\item intra: Within-cluster inertia (average variance within clusters)
\item inter: Between-cluster inertia (total - within)
\item pct_expliquee: Percentage of variance explained by clustering
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VAR_CAH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VAR_CAH-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR_CAH$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
